The script that ships the product is now enhanced.

+++ scripts/ShipProduct.sh
@@ -1,23 +1,56 @@
 #!/bin/bash
+# product release script
+# authored by Sébastien Boisvert
+
+echo "Welcome to Ray Technologies."
+echo "Today we ship a product."
+echo ""
+
+# you need 
+
+# - git
+# - mkdir (coreutils)
+# - bash
+# - xz
+# - tar
+# - gzip
+# - bzip2
+# - zip
+# - md5sum
+# - sha1sum
 
 release=$1
 
-RayGit=git://github.com/sebhtml/ray.git
-RayPlatformGit=git://github.com/sebhtml/RayPlatform.git
-ExampleGit=git://github.com/sebhtml/RayPlatform-example.git
+base=~/git-clones
 
-mkdir -p ~/ProductReleases/$release
-cd ~/ProductReleases/$release
+RayGit=$base/ray
+RayPlatformGit=$base/RayPlatform
+ExampleGit=$base/RayPlatform-example
+productName=Ray-$release
+productDirectory=~/ProductReleases/$productName
 
-git clone $RayGit Ray-$release
-cd Ray-$release
+if test -d $productDirectory
+then
+	echo "Product $productDirectory is already released !"
+	exit 1
+fi
+
+echo "Assembling product parts"
+
+mkdir -p $productDirectory
+cd $productDirectory
+
+echo "Cloning $RayGit"
+git clone $RayGit $productName &> /dev/null
+cd $productName
 git log|grep ^commit|head -n1|awk '{print $2}' > tag.txt
 echo https://github.com/sebhtml/ray > github.txt
 echo git://github.com/sebhtml/ray.git > git.txt
 rm -rf .git
 rm -rf RayPlatform
 
-git clone $RayPlatformGit
+echo "Cloning $RayPlatformGit"
+git clone $RayPlatformGit &> /dev/null
 cd RayPlatform
 git log|grep ^commit|head -n1|awk '{print $2}' > tag.txt
 echo https://github.com/sebhtml/RayPlatform > github.txt
@@ -25,7 +58,8 @@ echo git://github.com/sebhtml/RayPlatform.git > git.txt
 rm -rf .git
 cd ..
 
-git clone $ExampleGit
+echo "Cloning $ExampleGit"
+git clone $ExampleGit &> /dev/null
 cd RayPlatform-example
 git log|grep ^commit|head -n1|awk '{print $2}' > tag.txt
 echo https://github.com/sebhtml/RayPlatform-example > github.txt
@@ -34,10 +68,27 @@ rm -rf .git
 cd ..
 
 
+# create archives
+
+echo "Creating product releases."
 
 cd ..
-tar cjf Ray-$release.tar.bz2 Ray-$release
-tar czf Ray-$release.tar.gz Ray-$release
-zip -yr Ray-$release.zip Ray-$release
+echo ".tar.bz2"
+tar cjf $productName.tar.bz2 $productName &> /dev/null
+echo ".tar.gz"
+tar czf $productName.tar.gz $productName &> /dev/null
+echo ".zip"
+zip -yr $productName.zip $productName &> /dev/null
+echo ".tar.xz"
+tar cJf $productName.tar.xz $productName &> /dev/null
 
+rm -rf $productName
 
+# compute checksums
+
+for i in $(ls Ray-$release.*)
+do
+	sha1sum $i > $i.sha1sum.txt
+done
+
+echo "The product was shipped to $(pwd)"
